refactor(hermes): consolidate dashboard host validation
