+++ test/e2e/mock-parity.json
@@ -17,8 +17,7 @@
         "src/lib/onboard/docker-driver-gateway-env.test.ts",
         "src/lib/onboard/experimental/portable-host-preparation.test.ts",
         "src/lib/onboard/sandbox-prebuild.test.ts",
-        "test/agents/hermes/hermes-dashboard-external-host.test.ts",
-        "test/agents/hermes/hermes-dashboard-start.test.ts"
+        "test/agents/hermes/hermes-start.test.ts"
       ]
     },
     {
